// src/app/portal/moduls/scheduled-services/models/scheduledServices.ts
export class ScheduledServices {
    Code: string | null;
    ServiceType: string | null;
    IdReference: string | null;
    AgendaType: string | null;
    RDACode: string | null;
    ClinicCode: string | null;
    SpecialityCode: string | null;
    YearMonthDay: string | null;
    Time: string | null;
    Document: string | null;
    DocumentTypeUser: string | null;
    DocumentUser: string | null;
    Canceled: boolean | null;
    UserDocumentType: number | null;
    UserDocumentTypeStr: string | null;
    userDocument: number | null;
    UserName: string | null;
    ApplicantNameOne: string | null;
    ApplicantNameTwo: string | null;
    ApplicantLastNameOne: string | null;
    ApplicantLastNameTwo: string | null;
    ApplicantDocumentType: string | null;
    ApplicantDocumentTypeName: string | null;
    ApplicantDocument: string | null;
    ApplicantCellPhone: string | null;
    ApplicantMail: string | null;
  }
  